requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
Django==5.2.12
django-environ==0.13.0
django-storages==1.14.6
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-core==2.5.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
gunicorn==25.1.0
idna==3.11
packaging==26.0
pillow==12.1.1
proto-plus==1.27.1
protobuf==6.33.5
psycopg2-binary==2.9.11
qrcode==8.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.12.0
